Start refactoring `object()` into `create()`
