+++ tests/IntegrationTest/InheritanceTest.php
@@ -59,10 +59,7 @@ class InheritanceTest extends \PHPUnit_Framework_TestCase
         $builder->useAutowiring(true);
         $builder->useAnnotations(true);
         $containerAnnotations = $builder->build();
-        $containerAnnotations->set(
-            BaseClass::class,
-            \DI\object(SubClass::class)
-        );
+        $containerAnnotations->set(BaseClass::class, \DI\get(SubClass::class));
 
         // Test with a container using PHP configuration -> entries are different,
         // definitions shouldn't be shared between 2 different entries se we redefine all properties and methods
